Labs_05

+++ PyLesson_05/DiscountReceipt.py
@@ -1,11 +1,11 @@
 item1=input("Please enter item 1:")
-price1=int(input("Please enter the price of item 1:"))
+price1=float(input("Please enter the price of item 1:"))
 item2=input("Please enter item 2:")
-price2=int(input("Please enter the price of item 2:"))
+price2=float(input("Please enter the price of item 2:"))
 item3=input("Please enter item 3:")
-price3=int(input("Please enter the price of item 3:"))
+price3=float(input("Please enter the price of item 3:"))
 item4=input("Please enter item 4:")
-price4=int(input("Please enter the price of item 4:"))
+price4=float(input("Please enter the price of item 4:"))
